Change version type to long

+++ migration/src/main/java/br/com/bloder/migration/internal/detail/MigrationDetail.kt
@@ -3,10 +3,10 @@ package br.com.bloder.migration.internal.detail
 import br.com.bloder.migration.Migration
 
 /**
- * Created by bloder on 24/09/16.
+ * Better migrations properties
  */
 class MigrationDetail {
 
     var migration : Migration? = null
-    var version : Int = 0
+    var version : Long = 0
 }
